move schemas into package
