+++ source/databricks/package/schemas/output/aggregation_result_schema.py
