+++ source/databricks/package/steps/aggregation/aggregation_result_formatter.py
@@ -16,7 +16,7 @@ from pyspark.sql import DataFrame, SparkSession
 from pyspark.sql.functions import lit
 from geh_stream.shared.data_classes import Metadata
 from geh_stream.codelists import Colname
-from geh_stream.schemas.output import aggregation_result_schema
+from package.schemas.output import aggregation_result_schema
 
 
 def __add_missing_nullable_columns(result: DataFrame) -> DataFrame:
