+++ source/databricks/tests/integration/calculator/steps/aggregation/test_added_system_correction.py
@@ -22,7 +22,7 @@ from geh_stream.codelists import (
 from package.steps.aggregation import calculate_added_system_correction
 from geh_stream.codelists import Quality
 from geh_stream.shared.data_classes import Metadata
-from geh_stream.schemas.output import aggregation_result_schema
+from package.schemas.output import aggregation_result_schema
 from package.steps.aggregation.aggregation_result_formatter import (
     create_dataframe_from_aggregation_result_schema,
 )
