+++ source/databricks/tests/integration/calculator/steps/aggregation/test_flex_consumption_ga_brp_es.py
@@ -25,7 +25,7 @@ from geh_stream.codelists import (
     Quality,
 )
 from geh_stream.shared.data_classes import Metadata
-from geh_stream.schemas.output import aggregation_result_schema
+from package.schemas.output import aggregation_result_schema
 from pyspark.sql import DataFrame
 from pyspark.sql.types import StructType, StringType, DecimalType, TimestampType
 import pytest
