+++ source/databricks/tests/integration/calculator/steps/aggregation/test_hourly_production_ga_brp_es.py
@@ -20,7 +20,7 @@ from package.steps.aggregation import (
 )
 from geh_stream.codelists import MarketEvaluationPointType, ConnectionState, Quality
 from geh_stream.shared.data_classes import Metadata
-from geh_stream.schemas.output import aggregation_result_schema
+from package.schemas.output import aggregation_result_schema
 from pyspark.sql import DataFrame
 from pyspark.sql.types import StructType, StringType, DecimalType, TimestampType
 import pytest
