+++ source/databricks/tests/integration/calculator/steps/aggregation/test_net_exchange_grid_area_aggregation.py
@@ -19,7 +19,7 @@ from geh_stream.codelists import Colname, ResultKeyName
 from package.steps.aggregation import aggregate_net_exchange_per_ga
 from geh_stream.codelists import MarketEvaluationPointType, ConnectionState, Quality
 from geh_stream.shared.data_classes import Metadata
-from geh_stream.schemas.output import aggregation_result_schema
+from package.schemas.output import aggregation_result_schema
 from pyspark.sql import DataFrame
 import pyspark.sql.functions as F
 from pyspark.sql.types import StructType, StringType, DecimalType, TimestampType
