+++ source/databricks/tests/integration/calculator/steps/aggregation/test_net_exchange_neighbour.py
@@ -22,7 +22,7 @@ from package.steps.aggregation import (
 )
 from geh_stream.codelists import MarketEvaluationPointType, ConnectionState, Quality
 from geh_stream.shared.data_classes import Metadata
-from geh_stream.schemas.output import aggregation_result_schema
+from package.schemas.output import aggregation_result_schema
 from pyspark.sql.types import StructType, StringType, DecimalType, TimestampType
